# src/__test__/server/openapi/greeter.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "greeter.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "example.greeter.v1.services.Greeter"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/eqMetadata": {
      "post": {
        "summary": "客户端同时在metadata中和request中发送metadata，服务器会进行比较，如果相同返回 ok",
        "operationId": "Greeter_EqMetadata",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/example.greeter.v1.services.EqMetadataResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/google.rpc.Status"
            }
          }
        },
        "tags": [
          "example.greeter.v1.services.Greeter"
        ]
      }
    },
    "/v1/metadata": {
      "get": {
        "summary": "客户端在 request 中给出 metadata 服务端会在响应中设置 metadata，客户端可以进行比较",
        "operationId": "Greeter_Metadata",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/google.rpc.Status"
            }
          }
        },
        "tags": [
          "example.greeter.v1.services.Greeter"
        ]
      }
    },
    "/v1/sayHello/{name}": {
      "get": {
        "summary": "客户端 在 request 中 给出一个 name， 服务端会响应 hello 「name」",
        "operationId": "Greeter_SayHello",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/example.greeter.v1.services.HelloReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/google.rpc.Status"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "example.greeter.v1.services.Greeter"
        ]
      }
    },
    "/v1/status": {
      "get": {
        "summary": "客户端在 request 中给出一个状态码和错误信息，服务端会已对应的状态码进行响应",
        "operationId": "Greeter_Status",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/google.rpc.Status"
            }
          }
        },
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "errorMsg",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "example.greeter.v1.services.Greeter"
        ]
      }
    }
  },
  "definitions": {
    "example.greeter.v1.services.EqMetadataResponse": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean"
        }
      }
    },
    "example.greeter.v1.services.HelloReply": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "title": "The response message containing the greetings"
    },
    "google.protobuf.Any": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "google.rpc.Status": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/google.protobuf.Any"
          }
        }
      }
    }
  }
}